backend/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
requests
python-dotenv
cloudinary
flask-cors
google-auth
google-auth-oauthlib
google-auth-httplib2
